Move geometry endpoints to /molecules
Move the geometry endpoints to /molecules/{id}/geometries.

This is done because each geometry is a child of a molecule, and it
is reasonable to keep them together.
